pyproject.toml
[project]
name = "week-05-08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.26",
    "langchain-google-genai>=2.1.12",
    "pandas>=2.3.3",
    "pydantic==2.*",
    "python-dotenv>=1.2.1",
]
